requirements.txt
pygame
pyopengl
wikipedia
gTTS
playsound
pyperclip
